requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
msgspec==0.18.6
oauthlib==3.2.2
openai==1.42.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sniffio==1.3.1
SQLAlchemy==2.0.32
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3
